+++ src/refactoring/hooks/cart/useCart.ts
@@ -1,48 +1,32 @@
 // useCart.ts
 import { useState } from "react";
-import { CartItem, Coupon, Product } from "../../../types";
+import { CartItem, Product } from "../../../types";
 import {
   calculateCartTotal,
-  getRemainingStock,
+  addCartItemToCart,
+  isEmptyStock,
   updateCartItemQuantity,
+  removeCartItemFromCart,
 } from "../utils/cartUtils";
+import useSelectedCoupon from "./hooks/useSelectedCoupon";
 
 export const useCart = () => {
   const [cart, setCart] = useState<CartItem[]>([]);
-  const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
+  const { selectedCoupon, applyCoupon } = useSelectedCoupon();
 
   const addToCart = (product: Product) => {
-    const remainingStock = getRemainingStock(product, cart);
-    if (remainingStock <= 0) return;
-
-    setCart((prevCart) => {
-      const existingItem = prevCart.find(
-        (item) => item.product.id === product.id
-      );
-      if (existingItem) {
-        return prevCart.map((item) =>
-          item.product.id === product.id
-            ? { ...item, quantity: Math.min(item.quantity + 1, product.stock) }
-            : item
-        );
-      }
-      return [...prevCart, { product, quantity: 1 }];
-    });
+    if (isEmptyStock(product, cart)) return;
+    setCart((prevCart) => addCartItemToCart(prevCart, product));
   };
 
   const removeFromCart = (productId: string) => {
-    setCart((prevCart) =>
-      prevCart.filter((item) => item.product.id !== productId)
-    );
+    setCart((prevCart) => removeCartItemFromCart(prevCart, productId));
   };
 
   const updateQuantity = (productId: string, newQuantity: number) => {
     setCart((prevCart) =>
       updateCartItemQuantity(prevCart, productId, newQuantity)
     );
-  };
-  const applyCoupon = (coupon: Coupon) => {
-    setSelectedCoupon(coupon);
   };
 
   const calculateTotal = () => {
